requirements.txt
# NEURO COMMENT BOT — Dependencies
# ================================

# Telegram
telethon>=1.34.0
opentele>=1.15.1

# Web Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
aiofiles>=23.2.1

# Database
aiosqlite>=0.19.0

# OpenAI GPT
openai>=1.12.0

# Configuration
pyyaml>=6.0.1

# Proxy support
python-socks[asyncio]>=2.4.4
aiohttp>=3.9.3
aiohttp-socks>=0.8.4

# Utilities
python-dotenv>=1.0.0
colorama>=0.4.6

# File handling
python-multipart>=0.0.9
